+++ crates/swbus-core/Cargo.toml
@@ -41,15 +41,12 @@ futures-core.workspace = true
 
 # Internal dependencies
 swbus-proto.workspace = true
-swss-serde.workspace = true
-swss-common.workspace = true
 
 [dev-dependencies]
 pretty_assertions.workspace = true
 lazy_static.workspace = true
 # used in tests/
 swbus-edge.workspace = true
-swss-common-testing.workspace = true
 tracing-subscriber = { version = "0.3", features = ["env-filter", "serde"] }
 [build-dependencies]
 tonic-build.workspace = true
