+++ crates/swbus-core/src/mux/mod.rs
@@ -6,8 +6,8 @@ mod conn_worker;
 mod message_handler;
 mod multiplexer;
 pub mod nexthop;
-pub mod route_config;
 pub mod service;
+pub mod swbus_config;
 
 pub use conn::*;
 pub use conn_info::*;
@@ -16,4 +16,4 @@ pub use conn_worker::*;
 pub use message_handler::*;
 pub(crate) use multiplexer::*;
 pub(crate) use nexthop::*;
-pub(crate) use route_config::*;
+pub(crate) use swbus_config::*;
