+++ crates/swbusd/Cargo.toml
@@ -16,10 +16,23 @@ edition.workspace = true
 tokio.workspace = true
 tokio-stream.workspace = true
 tonic.workspace = true
-swbus-core.workspace = true
-sonic-common.workspace = true
 tracing.workspace = true
 clap = { version = "4.0", features = ["derive"] }
+serde.workspace = true
+serde_yaml.workspace = true
+serde_json.workspace = true
+tempfile.workspace = true
+anyhow.workspace = true
+
+# Internal dependencies
+swbus-core.workspace = true
+sonic-common.workspace = true
+swss-serde.workspace = true
+swss-common.workspace = true
+swbus-proto.workspace = true
+
+[dev-dependencies]
+swss-common-testing.workspace = true
 
 [lints]
 workspace = true
